+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 	OpenSeizureDetector Android App - Change Log
 	============================================
+    V5.0.8  - Fixed crash of advanced mode user interface (Issue #251)
+    V5.0.7  - Fixed potential crash after changing settings (Issue #244).
+            - Fixed issue where using the mute function on watch does not prevent fault pip sounds (Issue #247)
+    V5.0.6  - Disable 'Raise Alarm' button when using Network Data Source to prevent accidental alarm activations.
+	        - Reverted default alarmThreshold value back to 100 to be consistent with V4.3.1 and maintain good detection reliability by default.
+	        - Fixed background colour highlighting in Data Sharing screen.
+    V5.0.5  - Fixed potential race condition in LogManager that could prevent login and hang the UI
+    V5.0.4  - Added text to seizure detector settings screen to warn the user to disable basic mode if they want to change the seizure detector settings.
+    V5.0.3
+    V5.0.2  - Further improvements to user interface associated with authentication errors.
+            - Low bettery warning no only alerts if the device is running on battery, and not if it is charging.
+    V5.0.1  - Improved timeout error handling in authenticate activity
     V5.0.0  - Build for publication on Play Store  
             - Fixed issue with low battery warning not emitting fault sounds'
             - Corrected colour of OK/Cancel buttons in data sharing dialogs.